+++ NEWS.md
@@ -5,7 +5,9 @@
 
 * `use_readme_rmd()` once again creates a pre-commit git hook, to help keep `README.Rmd` and `README.md` in sync (@PeteHaitch @jennybc, #41).
 
-* `use_tidy_contributing` uses `tidy_contributing.md` template to make tidyverse-wide contributing guidelines (@batpigandme, #143)
+* `use_tidy_contributing` uses `tidy_contributing.md` template to make tidyverse-wide contributing guidelines (@batpigandme, #143).
+
+* `use_issues` uses `ISSUES.md` template to make `ISSUE_TEMPLATE.md` (@batpigandme).
 
 # usethis 1.1.0